Group Level Adjustment That Accounts for Propensity to Guess

Adjusts observed 1s based on propensity to guess (based on observed 0s) and item level \(\gamma\). You can also put in your best estimate of hidden knowledge behind don't know responses.

group_adj(pre = NULL, pst = NULL, gamma = NULL, dk = 0.03)

Arguments

pre

pre data frame. Required. Each vector within the data frame should only take values 0, 1, and 'd'.

pst

pst data frame. Required. Each vector within the data frame should only take values 0, 1, and 'd'.

gamma

probability of getting the right answer without knowledge

dk

Numeric. Between 0 and 1. Hidden knowledge behind don't know responses. Default is .03.

Value

nested list of pre and post adjusted responses, and adjusted learning estimates

Examples

pre_test_var <- data.frame(pre=c(1,0,0,1,"d","d",0,1,NA))
pst_test_var <- data.frame(pst=c(1,NA,1,"d",1,0,1,1,"d"))
gamma <- c(.25)
group_adj(pre_test_var, pst_test_var, gamma)

#> NAs will be converted to 0. MCAR is assumed.

#> $indiv
#> $indiv$pre_adj
#>         pre
#> 1 0.5555556
#> 2 0.0000000
#> 3 0.0000000
#> 4 0.5555556
#> 5 0.0300000
#> 6 0.0300000
#> 7 0.0000000
#> 8 0.5555556
#> 9 0.0000000
#> 
#> $indiv$pst_adj
#>         pst
#> 1 0.8666667
#> 2 0.0000000
#> 3 0.8666667
#> 4 0.0300000
#> 5 0.8666667
#> 6 0.0000000
#> 7 0.8666667
#> 8 0.8666667
#> 9 0.0300000
#> 
#> 
#> $learn
#>       pst 
#> 0.2962963 
#>